Cleanup: Moved all interfaces to their own namespace
