+++ examples/japierroruncaught/api.php
@@ -18,10 +18,10 @@
 
 namespace Docnet\JAPI\examples\japierrorcatching;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
 use Docnet\JAPI\error\JsonErrorHandler;
 use Docnet\JAPI\ErrorToException;
 use Docnet\JAPI\examples\controllers\Hello;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
 use Docnet\JAPI\JAPI;
 use Docnet\JAPI\middleware\CallStackFactory;
 use Docnet\JAPI\routing\Router;
