+++ src/interface/middleware/MiddlewareProviderInterface.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace Docnet\JAPI\middleware;
+namespace Docnet\JAPI\interface\middleware;
 
 /**
  * Middleware Provider Interface
