+++ src/middleware/CallStack.php
@@ -20,7 +20,9 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\middleware;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareProviderInterface;
 use gordonmcvey\httpsupport\request\RequestInterface;
 use gordonmcvey\httpsupport\response\ResponseInterface;
