+++ src/middleware/CallStackFactory.php
@@ -20,7 +20,8 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\middleware;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareProviderInterface;
 
 /**
  * Callstack Factory
