+++ src/middleware/Slot.php
@@ -20,7 +20,8 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\middleware;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareInterface;
 use gordonmcvey\httpsupport\request\RequestInterface;
 use gordonmcvey\httpsupport\response\ResponseInterface;
