+++ src/routing/SingleControllerStrategy.php
@@ -20,6 +20,8 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\routing;
 
+use Docnet\JAPI\interface\routing\RoutingStrategyInterface;
+
 /**
  * Single Controller Strategy
  *
