+++ tests/integration/middleware/CallStackTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\test\integration\middleware;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareProviderInterface;
 use Docnet\JAPI\middleware\CallStack;
-use Docnet\JAPI\middleware\MiddlewareInterface;
-use Docnet\JAPI\middleware\MiddlewareProviderInterface;
 use Docnet\JAPI\middleware\MiddlewareProviderTrait;
 use gordonmcvey\httpsupport\enum\statuscodes\SuccessCodes;
 use gordonmcvey\httpsupport\request\payload\ArrayPayloadHandler;
