+++ tests/unit/middleware/CallStackFactoryTest.php
@@ -20,10 +20,10 @@ declare(strict_types=1);
 
 namespace Docnet\JAPI\test\unit\middleware;
 
-use Docnet\JAPI\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\controller\RequestHandlerInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareInterface;
+use Docnet\JAPI\interface\middleware\MiddlewareProviderInterface;
 use Docnet\JAPI\middleware\CallStackFactory;
-use Docnet\JAPI\middleware\MiddlewareInterface;
-use Docnet\JAPI\middleware\MiddlewareProviderInterface;
 use gordonmcvey\httpsupport\request\RequestInterface;
 use gordonmcvey\httpsupport\response\ResponseInterface;
 use PHPUnit\Framework\Attributes\Test;
